Add: a test to capitalize a string Add: a function to capitalize a string Refactor: package.json to run all tests in the tests directory
