+++ package.json
@@ -7,7 +7,7 @@
     "test": "jest --coverage"
   },
   "scripts": {
-    "test": "jest --coverage"
+    "test": "jest --coverage --testPathPattern=tests/"
   },
   "repository": {
     "type": "git",
